These 4 functions implement a new approach to model specification that allows a user to specify a model in factored form when all state transitions are independent conditional on the current states and actions. The basic user interface is g2EV which operates much like g2P but the individual state transition functions are passed as a cell array and the state/action variables and shocks going into each function are passed as a cell array of lists.
g2EV first calls EVsetup to generate a sets of index vectors that are used in evaluated the EV function. These vectors indicate how to match variables at each evaluation step. EVsetup also generates a set of indices that can be used to obtain just the values of X necessary to compute the transition function. It uses these to evaluate the transition function and create CPT for the discretized state variables.

Lastly g2EV calls EVcreate, which return a function handle that can be passed to mdpsolve. The EV function makes successive calls to EVmergefunc, which is a MEX file that performs the specialized expansion and multiplication used by the EV function. EVmergefunc is a binary operator that takes two numerical arrays (X and Y) and two associated index vectors. The index vectors are the same length and indicate which columns of X and Y are used to create the associated column of the output. Although this operation could be implemented with bsxfun the MEX implementation is significantly faster.