Added the ability to pass a strategy as a function handle rather than a matrix. As a matrix the procedure uses table look up with nearest neighbor interpolation on an ns x da matrix. With a function the actions are determined using A=strategy(S) where strategy accepts a q by ds matrix of state values and returns a q by da matrix of action values.
Also added the ability to pass in a cell array (z) of random noise terms as inputs to the simulator. If this is empty then z is generated. On subsequent calls z can be passed back into dsim. If all other inputs are the same then dsim will return the same values. Passing in other values of s0, A or pval allows one to explore the effects of alternative starting values, strategies or parameters without the need to regenerate the underlying uniform, integer or normal random numbers.